% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GmeasureFunctions.R
\name{calcLogKernel}
\alias{calcLogKernel}
\title{Vector of log(k(x_i|nu_j,theta)) for a given nu_j and theta.}
\usage{
calcLogKernel(x, eta, theta)
}
\description{
Vector of log(k(x_i|nu_j,theta)) for a given nu_j and theta.
}
