% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GmeasureFunctions.R
\name{simFiniteGmeasure}
\alias{simFiniteGmeasure}
\title{Sim a finite Gmeasure}
\usage{
simFiniteGmeasure(par, w.delta = 0.1, r_x = 1, minGamma = 0)
}
\description{
simulate a cluster process on [0,1]x[0,1] with cluster centers
having a G measure with alpha < 0 and iid Gaussian dispersion
densities alpha < 0, beta > 0 and lambda > 0 are parameters of
the G measure theta is the spread of the isotropic Gaussian
dispersion density w.delta is the edge effect factor for
simulating the cluster centers
}
\examples{
GMT = simFiniteGmeasureThomas(alpha=1.5,beta = 0.1, lambda =5,theta = 0.03,givenLaten = TRUE)
plot(GMT$x,xlim = c(-0.2,1.2), ylim = c(-0.2,1.2),cex = 0.5)
points(GMT$c,cex = 0.7, col = 'red')
}
