% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GmeasureFunctions.R
\name{simPoisson}
\alias{simPoisson}
\title{t}
\usage{
simPoisson(mu, w.min, w.max)
}
\arguments{
\item{mu}{The constant intensity}

\item{w.min}{Lower bound of the sqare area}

\item{w.max}{Upper bound of the sqare area}
}
\value{
An array with the locations of the samples
}
\description{
Simulates a spatial Poisson process with constant intensity mu on a sqarea area
}
