% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GmeasureFunctions.R
\name{UpdateGprocessCenter}
\alias{UpdateGprocessCenter}
\title{Split: Sample one random center which is proposed for a split. 
A split is proposed by adding and subtracting the sampled vector to the original center, the angle of the
proposed vector is sampled from Von Mises distribution with mean equal an observed center of a cluster of the children. 
The mark to each proposed split is sampled by randomly portion out the mark of the original center, 
where one proportion is beta-distributed with parameters corresponding to the number of children in each cluster. 
The observed points previously assigned to the center is then divided randomly to new centers accordinlgy to the isotropic Gaussian dispersion density.
If the latent point has in the current less than 3 children, a more simple procedure is used.}
\usage{
UpdateGprocessCenter(D, P, u, r, accept, conf)
}
\arguments{
\item{D}{The data}

\item{P}{The current state of the MCMC}

\item{u}{Random uniform vector of numbers used in the MCMC for selctin split/merge/move}

\item{r}{Random uniform vector of numbers used for acceptance}

\item{accept}{Information regarding acceptance}

\item{lBext}{Boundary}

\item{kappaVonMises}{Paramter used in the Von Mises distribtuion for sampling angles for split}
}
\description{

}
