% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GmeasureFunctions.R
\name{updateM}
\alias{updateM}
\title{Update the parent label for the observed offsprings.}
\usage{
updateM(D, P)
}
\arguments{
\item{D}{The data}

\item{P}{The parameters}
}
\description{
Update the parent label for the observed offsprings
}
